udpates

+++ src/openattack.py
@@ -326,6 +326,8 @@ def parse_args():
 
 def openattack():
   import pickle
+  import os
+
   random.seed(OPTS.rng_seed)
   np.random.seed(OPTS.rng_seed)
   torch.manual_seed(OPTS.torch_seed)
@@ -341,9 +343,15 @@ def openattack():
   if OPTS.data_cache_dir:
     if not os.path.exists(OPTS.data_cache_dir):
         os.makedirs(OPTS.data_cache_dir)
-  word_mat = task_class.load_datasets(device, OPTS, word_mat_only=True)
-  pickle.dump(word_mat, open('./data/imdb_word_mat.pkl','wb'))
+
+  if not os.path.exists('./data/imdb_word_mat.pkl'):
+    word_mat = task_class.load_datasets(device, OPTS, word_mat_only=True)
+    pickle.dump(word_mat, open('./data/imdb_word_mat.pkl','wb'))
+  else:
+    word_mat = pickle.load(open('./data/imdb_word_mat.pkl','rb'))
+    
   model = task_class.load_model(word_mat, device, OPTS)
+
   print(model)
 
 if __name__ == '__main__':
